fix recent autotools complaints
error: 'pkglibdir' is not a legitimate directory for 'DATA'

According to 2022 autotools, pkglibdir ($libdir/$PACKAGE e.g. /lib/grub/) is not
the right place to put DATA into. We could put a LIB there instead (e.g. pkglibdir_LIB),
or, we can put DATA into the datadir, which is /usr/local/share. We choose the latter

Signed-off-by: Cristian Stoica <crististm@gmail.com>