fix objcopy problems - for real this time
The problem originates with a change in the linker prompted by a Fedora
requirement to support unique build IDs. This build ID is put into a
.note.gnu section that is both loadable and allocatable.
Because this section is indistinguishable from .text it is also copied
into a ROM binary into what become usually gigantic files.
Older compilers/linkers do not have this problem because the .note section
is not generated.

The solution used by other people is to remove this section when copying
object files.

https://lkml.org/lkml/2011/12/15/162
https://fedoraproject.org/wiki/Releases/FeatureBuildId

https://stackoverflow.com/questions/9401217/is-the-elf-notes-section-really-needed

Signed-off-by: Cristian Stoica <crististm@gmail.com>